refactor: post 관련 타입을 entities/post/model/types.ts로 이동
